+++ CHANGELOG.md
@@ -3,6 +3,22 @@
 
 This is the changelog for pdfnaut. Versions follow the scheme specified in the [Contributing Guide](./CONTRIBUTING.md#versioning). Dates are specified in YYYY-MM-DD format.
 
+## [unreleased]
+
+### Changes
+
+- No longer parse PDF header on every incremental update parse.
+- Emit a warning on encountering an unknown security handler.
+
+### Removals
+
+- `start_xref` parameter in `PdfParser.parse`. The `parse()` function was rewritten so as to no longe use this parameter.
+- `cos.helpers` module. Removed due to circular import issues. Its functions were moved to other modules: `deref`, `into_bytes`, and `is_null_like` were moved to the `cos.objects.base` module; `ensure` is now an internal function.
+
+### Fixes
+
+- Fix wrong type hint of the `ids` parameter of `StandardSecurityHandler`.
+
 ## [0.13.0] (2026-08-07)
 
 ### Additions
@@ -324,7 +340,7 @@ This is the changelog for pdfnaut. Versions follow the scheme specified in the [
 
 The first release of pdfnaut. :tada:
 
-<!-- [unreleased]: https://github.com/aescarias/pdfnaut/compare/v0.13...HEAD -->
+[unreleased]: https://github.com/aescarias/pdfnaut/compare/v0.13...HEAD
 [0.13.0]: https://github.com/aescarias/pdfnaut/compare/v0.12...v0.13
 [0.12.0]: https://github.com/aescarias/pdfnaut/compare/v0.11.1...v0.12
 [0.11.1]: https://github.com/aescarias/pdfnaut/compare/v0.11...v0.11.1